chatbot/requirements.txt
streamlit
requests
beautifulsoup4
google-generativeai
PyPDF2
sentence-transformers
faiss-cpu
numpy
